Jeez, I did this another way that just kept breaking if( you tried to change last or first name)
Fill in the object constructor with the methods specified in the tests.

Those methods are getFirstName(), getLastName(), getFullName(),
setFirstName(first), setLastName(last), and setFullName(firstAndLast).

All functions that take an argument have an arity of 1, and the
argument will be a string.

These methods must be the only available means for interacting with the
object.

Remember to use RSAP if you get stuck. Try to pair program. Write your
own code.

Here are some helpful links:

Closures
Details of the Object Model